Changed file structure,default size is 1000
